split up step
