+++ service/src/main/java/bio/terra/workspace/service/resource/controlled/flight/clone/bucket/DeleteStorageTransferServiceJobStep.java
@@ -1,0 +1,36 @@
+package bio.terra.workspace.service.resource.controlled.flight.clone.bucket;
+
+import bio.terra.stairway.FlightContext;
+import bio.terra.stairway.Step;
+import bio.terra.stairway.StepResult;
+import bio.terra.stairway.exception.RetryException;
+import bio.terra.workspace.service.resource.model.CloningInstructions;
+import bio.terra.workspace.service.workspace.flight.WorkspaceFlightMapKeys.ControlledResourceKeys;
+
+/**
+ * Delete a Storage Transfer Service job, which has already had an operation run to completion or
+ * failure.
+ */
+public class DeleteStorageTransferServiceJobStep implements Step {
+
+  @Override
+  public StepResult doStep(FlightContext flightContext)
+      throws InterruptedException, RetryException {
+    final CloningInstructions effectiveCloningInstructions =
+        flightContext
+            .getWorkingMap()
+            .get(ControlledResourceKeys.CLONING_INSTRUCTIONS, CloningInstructions.class);
+    // This step is only run for full resource clones
+    if (CloningInstructions.COPY_RESOURCE != effectiveCloningInstructions) {
+      return StepResult.getStepResultSuccess();
+    } else {
+      return StorageTransferServiceUtils.deleteTransferJobStepImpl(flightContext);
+    }
+  }
+
+  // Nothing to undo
+  @Override
+  public StepResult undoStep(FlightContext flightContext) throws InterruptedException {
+    return StepResult.getStepResultSuccess();
+  }
+}
