+++ service/src/main/java/bio/terra/workspace/service/workspace/flight/WorkspaceFlightMapKeys.java
@@ -51,6 +51,7 @@ public final class WorkspaceFlightMapKeys {
     public static final String DESTINATION_BUCKET_CLONE_INPUTS = "destinationBucketCloneInputs";
     public static final String CONTROL_PLANE_PROJECT_ID = "controlPlaneProjectId";
     public static final String STORAGE_TRANSFER_SERVICE_SA_EMAIL = "storageTransferServiceSAEmail";
+    public static final String STORAGE_TRANSFER_JOB_NAME = "storageTransferJobName";
   }
 
   /** Common resource keys */
